+++ Simple.BTC-RPC/Simple.BTC/Models/Blcokchain/GetChainStates.cs
@@ -2,17 +2,17 @@
 
 public class GetChainStates_Result
 {
-    public int headers { get; set; }
+    public long headers { get; set; }
     public Chainstate[] chainstates { get; set; }
 }
 
 public class Chainstate
 {
-    public int blocks { get; set; }
+    public long blocks { get; set; }
     public string bestblockhash { get; set; }
-    public float difficulty { get; set; }
-    public float verificationprogress { get; set; }
-    public int coins_db_cache_bytes { get; set; }
+    public decimal difficulty { get; set; }
+    public decimal verificationprogress { get; set; }
+    public long coins_db_cache_bytes { get; set; }
     public long coins_tip_cache_bytes { get; set; }
     public bool validated { get; set; }
 }
